refactor(Reactions): move svgs into image folder
